pyproject.toml
[project]
name = "mytorch"
version = "0.1.0"
description = "MyTorch v1 — minimal DNN/CNN/RNN framework with autograd, NumPy/CuPy backends"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.3.4",
    "cupy-cuda12x>=13.6.0",
]

[tool.setuptools.packages.find]
include = ["mytorch*"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
